pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chargeflow"
version = "0.1.0"
description = "AI-assisted EV charging reliability and analytics platform"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "pandas>=2.2,<3.0",
  "numpy>=1.26,<3.0",
  "scikit-learn>=1.5,<2.0",
  "joblib>=1.4,<2.0",
  "groq>=0.20,<1.0",
  "pyyaml>=6.0,<7.0",
  "requests>=2.32,<3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1,<9.0",
]

[project.scripts]
chargeflow-ingest = "src.ingestion.cli:main"
chargeflow-synthetic = "src.synthetic.cli:main"
chargeflow-warehouse = "src.warehouse.cli:main"
chargeflow-ml = "src.ml.cli:main"
chargeflow-rag = "src.rag.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
